+++ TimeThreads/ChangeLog.md
@@ -14,7 +14,9 @@
 - [x] Refrained Model part: to fit the needs
 - [x] Implement subTask addition and edition
 - [x] Debug case: newly added task cannot set label at the first time
-- [ ] Refraine Base struct to support further development and modifications
+- [x] Refraine Base struct to support further development and modifications
+- [x] organized navigation views 
+- [x] solve the update problem of targetList, taskList and subTaskList
 - [ ] drag and drop of the subTasks
 - [ ] deleting subTask
 - [ ] Interaction of tasks: new, move, delete